[IMP] add method get_instance_data to notification
This method is to be implemented by every child of
the notification class. It's porpuse is to add each
notification type specific data to be sent.

Also the get_data method was move to the parent class notification
this will allow general behavior to be defined. In this case,
for some reason, the field delayed_option should no be sent if not set.